requirements.txt
langchain
langchain_community
langchain_core
langchain_huggingface
faiss-cpu
pypdf
huggingface_hub
flask
python-dotenv
sentence-transformers
langchain_groq